Device update after first run now possible.
Change log for v0.2.5 2020-01-11
Added the 'create_device' flag to allow new devices to be created. Required on first run for plugin. Domoticz log will show device creation errors on subsequent plugin startup if this is left on, but this should not cause any issues.  Set to True if you wish to add a device.  Set to False once all devices are created.  Restart the plugin Hardware each time it is changed.

Note that you must delete a device if you wish to change its type or subtype or any other parameters.  You don't need to delete all of the Devices created by the plugin.

Updated some of the Domoticz.Debug(...) lines for consistency.